+++ backend-rust/src/migrations/m0023_add_init_parameter.rs
@@ -1,17 +1,15 @@
 use super::{SchemaVersion, Transaction};
 use crate::transaction_event::events_from_summary;
-use anyhow::Context;
+use anyhow::{anyhow, Context};
 use async_graphql::futures_util::StreamExt;
 use chrono::{DateTime, Utc};
 use concordium_rust_sdk::{
-    types::{
-        AbsoluteBlockHeight, AccountTransactionEffects, BlockItemSummaryDetails, TransactionType,
-    },
+    types::{AbsoluteBlockHeight, AccountTransactionEffects, BlockItemSummaryDetails},
     v2::{self},
 };
 
-/// Performs a migration that alters the events of transactions being a transfer
-/// with schedule
+/// Performs a migration that alters the events types for transactions of type
+/// `Account::InitializeSmartContractInstance`
 pub async fn run(
     tx: &mut Transaction,
     endpoints: &[v2::Endpoint],
@@ -28,9 +26,10 @@ pub async fn run(
         "
             SELECT
                 block_height,
-                (SELECT slot_time FROM blocks WHERE height = block_height) as block_slot_time
+                (SELECT slot_time FROM blocks WHERE height = block_height) as block_slot_time,
+                COUNT(*) AS update_count
             FROM transactions
-            WHERE type_account IN ('TransferWithSchedule', 'TransferWithScheduleWithMemo')
+            WHERE type = 'Account' AND type_account = 'InitializeSmartContractInstance' AND success
             GROUP BY block_height
             ",
     )
@@ -39,6 +38,7 @@ pub async fn run(
 
     for row in rows {
         let height: i64 = sqlx::Row::try_get(&row, "block_height")?;
+        let mut expected_rows_to_be_affected_count: i64 = sqlx::Row::try_get(&row, "update_count")?;
         let block_slot_time: DateTime<Utc> = sqlx::Row::try_get(&row, "block_slot_time")?;
         let mut block_summary = client
             .get_block_transaction_events(AbsoluteBlockHeight {
@@ -47,27 +47,18 @@ pub async fn run(
             .await?
             .response;
         while let Some(summary) = block_summary.next().await.transpose()? {
-            let BlockItemSummaryDetails::AccountTransaction(update) = &summary.details else {
+            let BlockItemSummaryDetails::AccountTransaction(details) = &summary.details else {
                 continue;
             };
-            if !matches!(
-                update.transaction_type(),
-                Some(
-                    TransactionType::TransferWithSchedule
-                        | TransactionType::TransferWithScheduleAndMemo
-                )
-            ) {
+            let AccountTransactionEffects::ContractInitialized {
+                data: _,
+            } = details.effects
+            else {
                 continue;
-            }
-            if !matches!(
-                update.effects,
-                AccountTransactionEffects::TransferredWithSchedule { .. }
-                    | AccountTransactionEffects::TransferredWithScheduleAndMemo { .. }
-            ) {
-                continue;
-            }
+            };
 
             let events = events_from_summary(summary.details, block_slot_time)?;
+            let json = serde_json::to_value(events)?;
             let hash = summary.hash.to_string();
             sqlx::query(
                 "
@@ -76,10 +67,19 @@ pub async fn run(
                 WHERE hash = $2;
             ",
             )
-            .bind(serde_json::to_value(events)?)
+            .bind(json)
             .bind(hash)
             .execute(tx.as_mut())
             .await?;
+            expected_rows_to_be_affected_count -= 1;
+        }
+        let is_misalignment_between_expected_and_actual_rows_updated_count =
+            expected_rows_to_be_affected_count != 0;
+        if is_misalignment_between_expected_and_actual_rows_updated_count {
+            return Err(anyhow!(
+                "There is a misalignment between expected number of rows to be updated and actual \
+                 number of rows to have become updated"
+            ));
         }
     }
     Ok(next_schema_version)