+++ frontend/src/queries/useTransactionQuery.ts
@@ -198,6 +198,10 @@ __typename
 			chainUpdate
 			finalizationProof
 		}
+		... on ValidatorScoreParametersUpdate {
+			maximumMissedRounds
+			__typename
+		}
 		...on MicroCcdPerEuroChainUpdatePayload {
 			exchangeRate {
 				denominator
@@ -491,10 +495,6 @@ reason {
 		bakerId
 		__typename
   }
-  ... on ValidatorScoreParametersUpdate {
-		maximumMissedRounds
-		__typename
-  }
   ... on StakeOverMaximumThresholdForPool {
 		__typename
   }
@@ -639,6 +639,7 @@ export const useTransactionQuery = ({
 	const identifier = id?.value ? { id: id.value } : { hash: hash?.value }
 
 	const { data, fetching, error } = useQuery({
+		context: { url: useRuntimeConfig().public.apiUrlRust },
 		query,
 		requestPolicy: 'cache-first',
 		variables: {