+++ frontend/src/queries/useTransactionReleaseSchedule.ts
@@ -56,6 +56,7 @@ export const useTransactionReleaseSchedule = (
 	schedulePaging: QueryVariables
 ) => {
 	const { data } = useQuery({
+		context: { url: useRuntimeConfig().public.apiUrlRust },
 		query: TransactionReleaseScheduleQuery,
 		requestPolicy: 'cache-first',
 		variables: {